ch 11: resolver
the resolver fits in like so:

scanner -> parser -> resolver -> interpreter

it is needed because we need to know a block-by-block hop distance
of the environment definitions. Each line is resolved in this case so
we are actually storing a line-by-line hop distance.
in ex ch11, the lookup for showA returns "0" because it was defined at the
global scale (not found in the locals definition).

In a more complex example, we'd need to check hops of the blocks or
assignments to check what scope the variable being accessed in is from.